test: simplify validation coverage
