+++ packages/compiler/package.json
@@ -569,7 +569,7 @@
     "unicode:generate": "node scripts/generate-unicode-tables.mjs",
     "unicode:vendor": "node scripts/vendor-unicode-data.mjs",
     "unicode:vendor:check": "node scripts/vendor-unicode-data.mjs --check",
-    "test": "pnpm unicode:check && pnpm stdlib:check && pnpm toolchain:generate && pnpm documentation:policy && pnpm documentation:check && pnpm documentation:examples && pnpm test:parallel && pnpm test:native-acceptance",
+    "test": "pnpm unicode:check && pnpm stdlib:check && pnpm toolchain:generate && pnpm documentation:check && pnpm documentation:examples && pnpm test:parallel && pnpm test:native-acceptance",
     "test:native-acceptance": "vitest run test/DriverNativeAcceptance.test.ts --passWithNoTests --maxWorkers=1",
     "test:parallel": "vitest run --passWithNoTests --exclude test/DriverNativeAcceptance.test.ts",
     "typecheck": "pnpm unicode:check && pnpm stdlib:check && pnpm toolchain:generate && tsc -p tsconfig.json --noEmit && tsc -p tsconfig.test.json",
@@ -577,7 +577,6 @@
     "documentation:check": "node scripts/generate-documentation.mjs --check",
     "documentation:examples": "pnpm --filter @silklang/cli doctest:stdlib",
     "documentation:inventory": "node scripts/inventory-stdlib-documentation.mjs",
-    "documentation:policy": "node scripts/check-stdlib-documentation.mjs",
     "test:foreign-contracts": "node conformance/foreign-contracts/run.mjs",
     "test:native-boundary": "node conformance/native-boundary/run.mjs",
     "test:llvm-helpers": "node conformance/llvm-helpers/run.mjs",
